위원회를 삭제할때, 소속의원이 존재 할 경우 의원을 먼저 삭제하라는 알림을 띄워줍니다

+++ app/controllers/admin/committees_controller.rb
@@ -30,8 +30,13 @@ module Admin
     end
 
     def destroy
-      errors_to_flash(@committee) unless @committee.destroy
-      redirect_to admin_committees_url
+      if @committee.congressmen.present?
+        flash[:notice] = t('messages.has_congressman')
+        redirect_to admin_committees_url
+      else
+        errors_to_flash(@committee) unless @committee.destroy
+        redirect_to admin_committees_url
+      end
     end
 
     private
